+++ components/map-side-panel.tsx
@@ -31,7 +31,6 @@ type MapSidePanelProps = {
   selectedCountry: MapCountryMetricRow | null;
   countryDataReady: boolean;
   detailSelectionActive: boolean;
-  detailAccessLocked: boolean;
   selectedPublisher: MapPublisherMetricRow | null;
   selectedPublisherWindowMetrics: MapPublisherMetricRow['windows'][MapMetricWindow] | null;
   totals: MapCountryMetricsResponse['totals'] | null;
@@ -88,7 +87,6 @@ export function MapSidePanel({
   selectedCountry,
   countryDataReady,
   detailSelectionActive,
-  detailAccessLocked,
   selectedPublisher,
   selectedPublisherWindowMetrics,
   totals,